server/server_microservices_9/requirements.txt
flask>=2.0.0
flask-cors>=3.0.0
onnxruntime-gpu>=1.16.0
numpy>=1.21.0
opencv-python>=4.5.0

